+++ bin/cli.js
@@ -14,13 +14,15 @@ var argv = minimist(process.argv.slice(2), {
     'version'
   ],
   alias: {
-    h: 'help',
-    v: 'version',
+    b: 'config-basedir',
+    c: 'config',
     d: 'diff',
+    h: 'help',
+    i: 'ignore-path',
+    id: 'ignore-disables',
     l: 'list',
     R: 'recursive',
-    b: 'config-basedir',
-    c: 'config'
+    v: 'version',
   }
 })
 
@@ -38,14 +40,16 @@ if (argv.h) {
   console.log('')
   console.log('Options:')
   console.log('')
-  console.log('  -d, --diff             output diff against original file')
-  console.log('  -l, --list             format list of space seperated files in place')
-  console.log('  -R, --recursive        format files recursively')
-  console.log('  -c, --config           path to a specific configuration file (JSON, YAML, or CommonJS)')
+  console.log('  -d, --diff             Output diff against original file')
+  console.log('  -l, --list             Format list of space seperated files in place')
+  console.log('  -R, --recursive        Format files recursively')
+  console.log('  -c, --config           Path to a specific configuration file (JSON, YAML, or CommonJS)')
   console.log('  -b, --config-basedir   path to the directory that relative paths defining "extends"')
-  console.log('  -v, --version          output the version number')
-  console.log('  -h, --help             output usage information')
-  process.exit()
+  console.log('  -v, --version          Output the version number')
+  console.log('  -h, --help             Output usage information')
+  console.log('  -i, --ignore-path      Path to a file containing patterns that describe files to ignore.')
+  console.log('  -id --ignore-disables  Ignore disables')
+  console.log('  --stdin-filename       A filename to assign stdin input.')
 }
 
 
@@ -58,6 +62,14 @@ if (argv.b) {
   options.configBasedir = (path.isAbsolute(argv.b))
     ? argv.b
     : path.resolve(process.cwd(), argv.b)
+}
+
+if (argv.i) {
+  options.ignorePath = argv.i
+}
+
+if (argv.id) {
+  options.ignoreDisables = argv.id
 }
 
 if (argv.l) {
@@ -96,9 +108,14 @@ if (argv.l) {
     processMultipleFiles(files)
   })
 } else {
+
   stdin(function (css) {
+    options.codeFilename = argv['stdin-filename']
     postcss([stylefmt(options)])
-      .process(css, { syntax: scss })
+      .process(css, {
+        from: options.codeFilename,
+        syntax: scss
+      })
       .then(function (result) {
         process.stdout.write(result.css)
       })