+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## HEAD
+
+- Added: `--stdin-filename` option to CLI. same as [stylelint CLI](https://github.com/stylelint/stylelint/blob/master/src/cli.js#L75)
+- Added: `--ignore-path` option to CLI. same as [stylelint CLI](https://github.com/stylelint/stylelint/blob/master/src/cli.js#L63)
+- Added: `ignorePath` option to Node.js API. same as [stylelint API](http://stylelint.io/user-guide/node-api/#ignorepath)
+- Added: `ignoreFiles` option to configuration file. same as [stylelint configuration](http://stylelint.io/user-guide/configuration/#ignorefiles)
+- Added: `--ignore-disables` option to CLI.
+- Added: `ignoreDisables` option to Node.js API.
+
 ## v4.4.0
 
 - Fixed [Same issue in stylelint #218](https://github.com/stylelint/stylelint/issues/1881) use `stylelint` to load configuration file.